Move RendererMetrics into its own file
