+++ cpp/include/helpers/Thread.h
@@ -50,24 +50,4 @@ namespace rive_android
 	void setAffinity(int32_t cpu);
 
 	void setAffinity(Affinity affinity);
-
-	// This is a minimal demonstration of a thread manager that uses pthreads.
-	// It is not intended for production use!
-	struct ThreadManager
-	{
-		std::mutex threadMapMutex;
-		std::unordered_map<SwappyThreadId, pthread_t> threads;
-		ThreadManager() {}
-		ThreadManager(const ThreadManager&) = delete;
-		static std::atomic<SwappyThreadId> nextId;
-
-	public:
-		static ThreadManager& Instance();
-		int Start(SwappyThreadId* thread_id,
-		          void* (*thread_func)(void*),
-		          void* user_data);
-		void Join(SwappyThreadId thread_id);
-		bool Joinable(SwappyThreadId thread_id);
-	};
-
 } // namespace rive_android
